order succes

+++ app/src/main/java/com/focx/domain/entity/Order.kt
@@ -54,5 +54,5 @@ data class CreateOrder(
 @Serializable
 data class OrderPayment(
     val productId: ULong,
-    val quantity: UInt,
+    val quantity: ULong,
 )
